requirements.txt
flask==2.0.1
mlflow==1.16.0 
numpy==1.20.2
pandas==1.2.4
scikit-learn==0.24.1
lightgbm==3.2.0
click==7.1.2
python-dotenv==0.17.0
verispy==0.1.13
boto3==1.17.58
connexion==2.7.0
swagger-ui-bundle==0.0.8


